Added method to switch group and improved "toImageJ" method for some rotated shapes.

+++ src/test/java/fr/igred/omero/repository/DatasetTest.java
@@ -119,7 +119,7 @@ public class DatasetTest extends UserTest {
 
         assertEquals("TestDataset", dataset.getName());
         assertEquals("description", dataset.getDescription());
-        assertEquals(1L, dataset.getId().longValue());
+        assertEquals(1L, dataset.getId());
     }
